metoffice/app/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
psycopg2-binary==2.9.5
python-decouple==3.7
requests==2.28.1
urllib3==1.26.13